Change to a class.
